+++ lesson19/my-app/src/app/recently-added/recently-added.component.html
@@ -8,8 +8,7 @@
     <p>responsible - ответственный</p>
   </div>
   <div>
-    <button mat-button (click)="increase($event)">Add New</button>
-    <span>Количество кликов: {{count}}</span>
+    <button mat-button>Add New</button>
   </div>
   <div>
     <input type="text" #textToTranslate/>
